Run-length encode the world.
When writing the world, instead of writing successive location lines
with the same info, instead write the number of identical locations
followed by their info. This removes a bunch of duplicated lines from
the world representation, making it smaller to store and a lot faster
to read and write.

For example, seed 0 (of minima, so including animals) was 1.7MB before
this change and is 765KB after.


Fixes issue 113.